// app/models/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ConnectionCursor: any;
  ConnectionLimitInt: any;
  Date: any;
  DateTime: any;
  Email: any;
  JSONObject: any;
};
// ...
export type CatalogProduct = CatalogProductOrVariant &
  Node & {
    __typename?: 'CatalogProduct';
    /** The CatalogProduct ID. Do not assume that this is the same as the related product ID. See `productId` for that. */
    _id: Scalars['ID'];
    /** The product barcode value, if it has one */
    barcode?: Maybe<Scalars['String']>;
    /** The date and time at which this CatalogProduct was created, which is when the related product was first published */
    createdAt: Scalars['DateTime'];
    /** The full product description, which may have newline characters in it */
    description?: Maybe<Scalars['String']>;
    /** The height of the product, if it has physical dimensions */
    height?: Maybe<Scalars['Float']>;
    /**
     * True if every purchasable variant of this product is sold out but allows backorders. A storefront UI may use this
     * to decide to show a "Backordered" indicator.
     */
    isBackorder: Scalars['Boolean'];
    /** True if this product has been deleted. Typically, deleted products are not returned in queries. */
    isDeleted: Scalars['Boolean'];
    /**
     * True if at least one purchasable variant of this product has a low quantity in stock. A storefront UI may use this
     * to decide to show a "Low Quantity" indicator.
     */
    isLowQuantity: Scalars['Boolean'];
    /**
     * True if every purchasable variant of this product is sold out. A storefront UI may use this
     * to decide to show a "Sold Out" indicator when `isBackorder` is not also true.
     */
    isSoldOut: Scalars['Boolean'];
    /** True if this product should be shown to shoppers. Typically, non-visible products are not returned in queries. */
    isVisible: Scalars['Boolean'];
    /** The length of the product, if it has physical dimensions */
    length?: Maybe<Scalars['Float']>;
    /** All media for this product and its variants */
    media?: Maybe<Array<Maybe<ImageInfo>>>;
    /** The product description to use for page `description` meta element in HTML */
    metaDescription?: Maybe<Scalars['String']>;
    /** Arbitrary additional metadata about this product */
    metafields?: Maybe<Array<Maybe<Metafield>>>;
    /** The minimum quantity that must be added to a cart */
    minOrderQuantity?: Maybe<Scalars['Int']>;
    /** The country of origin */
    originCountry?: Maybe<Scalars['String']>;
    /** Subtitle */
    pageTitle?: Maybe<Scalars['String']>;
    /** Dimensions and other information about the containers in which this product will be shipped */
    parcel?: Maybe<ShippingParcel>;
    /** Price and related information, per currency */
    pricing: Array<Maybe<ProductPricingInfo>>;
    /** The primary image */
    primaryImage?: Maybe<ImageInfo>;
    /** The related Product ID */
    productId: Scalars['ID'];
    /** An arbitrary product type value, such as from an external system */
    productType?: Maybe<Scalars['String']>;
    /** The shop to which this product belongs */
    shop: Shop;
    /** A stock keeping unit (SKU) identifier for this product */
    sku?: Maybe<Scalars['String']>;
    /** A URL-safe and human-readable string that uniquely identifies this product */
    slug?: Maybe<Scalars['String']>;
    /** Holds metadata specific to a specific social network service */
    socialMetadata?: Maybe<Array<Maybe<SocialMetadata>>>;
    /** When a shopper purchases this product, what types of fulfillment can they choose from? */
    supportedFulfillmentTypes: Array<Maybe<FulfillmentType>>;
    /** The list of tag IDs that have been applied to this product */
    tagIds?: Maybe<Array<Maybe<Scalars['ID']>>>;
    /** The list of tags that have been applied to this product */
    tags?: Maybe<TagConnection>;
    /** Product title */
    title?: Maybe<Scalars['String']>;
    /** The date and time at which this CatalogProduct was last updated, which is when the related product was most recently published */
    updatedAt: Scalars['DateTime'];
    /** A flat list of all variants for this product */
    variants?: Maybe<Array<Maybe<CatalogProductVariant>>>;
    /** The product vendor or manufacturer, for display */
    vendor?: Maybe<Scalars['String']>;
    /** The weight of the product on Earth, if it has physical dimensions */
    weight?: Maybe<Scalars['Float']>;
    /** The width of the product, if it has physical dimensions */
    width?: Maybe<Scalars['Float']>;
  };
// ...
export type CatalogProductOrVariant = {
  /** The product barcode value, if it has one */
  barcode?: Maybe<Scalars['String']>;
  /** The date and time at which this CatalogProduct was created, which is when the related product was first published */
  createdAt?: Maybe<Scalars['DateTime']>;
  /** The height of the product, if it has physical dimensions */
  height?: Maybe<Scalars['Float']>;
  /** The length of the product, if it has physical dimensions */
  length?: Maybe<Scalars['Float']>;
  /** Arbitrary additional metadata about this product */
  metafields?: Maybe<Array<Maybe<Metafield>>>;
  /** The minimum quantity that must be added to a cart */
  minOrderQuantity?: Maybe<Scalars['Int']>;
  /** The country of origin */
  originCountry?: Maybe<Scalars['String']>;
  /** The shop to which this product belongs */
  shop: Shop;
  /** A stock keeping unit (SKU) identifier for this product */
  sku?: Maybe<Scalars['String']>;
  /** Product or variant title */
  title?: Maybe<Scalars['String']>;
  /** The date and time at which this CatalogProduct was last updated, which is when the related product was most recently published */
  updatedAt?: Maybe<Scalars['DateTime']>;
  /** The weight of the product on Earth, if it has physical dimensions */
  weight?: Maybe<Scalars['Float']>;
  /** The width of the product, if it has physical dimensions */
  width?: Maybe<Scalars['Float']>;
};
// ...
export type Deletable = {
  /**
   * If true, this object should be considered deleted. Soft deleted objects are not
   * returned in query results unless you explicitly ask for them.
   */
  isDeleted: Scalars['Boolean'];
};
// ...
export enum DistanceUnit {
  /** Centimeter */
  Cm = 'cm',
  /** Foot */
  Ft = 'ft',
  /** Inch */
  In = 'in',
}
// ...
export enum FulfillmentType {
  /** An order will be fulfilled digitally, such as by sending a download link */
  Digital = 'digital',
  /** An order will be fulfilled by the customer picking it up */
  Pickup = 'pickup',
  /** An order will be fulfilled by the seller shipping it to the customer */
  Shipping = 'shipping',
}
// ...
export type ImageInfo = {
  __typename?: 'ImageInfo';
  /** A list of URLs for various size files of this image */
  URLs?: Maybe<ImageSizes>;
  /** ID */
  _id?: Maybe<Scalars['ID']>;
  /**
   * Sort by priority ascending when displaying more than one image for a product in a user interface.
   * This is an integer with 1 being the first / highest priority image.
   */
  priority?: Maybe<Scalars['Int']>;
  /** The related product ID */
  productId?: Maybe<Scalars['ID']>;
  /** The related variant ID, if linked with a particular variant */
  variantId?: Maybe<Scalars['ID']>;
};
// ...
export enum MassUnit {
  /** Gram */
  G = 'g',
  /** Kilogram */
  Kg = 'kg',
  /** Pound */
  Lb = 'lb',
  /** Ounce */
  Oz = 'oz',
}
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** Field description */
  description?: Maybe<Scalars['String']>;
  /** Field key */
  key?: Maybe<Scalars['String']>;
  /** Field namespace */
  namespace?: Maybe<Scalars['String']>;
  /** Field scope */
  scope?: Maybe<Scalars['String']>;
  /** Field value */
  value?: Maybe<Scalars['String']>;
  /** Field value type */
  valueType?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  /** The ID of the object */
  _id: Scalars['ID'];
};
// ...
export type ProductPricingInfo = {
  __typename?: 'ProductPricingInfo';
  /**
   * A comparison price value, usually MSRP. If `price` is null, this will also be null. That is,
   * only purchasable variants will have a `compareAtPrice`.
   */
  compareAtPrice?: Maybe<Money>;
  /** The code for the currency these pricing details applies to */
  currency: Currency;
  /** Pricing converted to specified currency */
  currencyExchangePricing?: Maybe<CurrencyExchangeProductPricingInfo>;
  /**
   * UI should display this price. If a product has multiple potential prices depending on selected
   * variants and options, then this is a price range string such as "$3.95 - $6.99". It includes the currency
   * symbols.
   */
  displayPrice: Scalars['String'];
  /** The price of the most expensive possible variant+option combination */
  maxPrice: Scalars['Float'];
  /** The price of the least expensive possible variant+option combination */
  minPrice: Scalars['Float'];
  /**
   * For variants with no options and for options, this will always be set to a price. For variants
   * with options and products, this will be `null`. There must be a price for a variant to be
   * added to a cart or purchased. Otherwise you would instead add one of its child options to a cart.
   */
  price?: Maybe<Scalars['Float']>;
};
// ...
export type ShippingParcel = {
  __typename?: 'ShippingParcel';
  /** Containers */
  containers?: Maybe<Scalars['String']>;
  /** Distance unit */
  distanceUnit?: Maybe<DistanceUnit>;
  /** Height */
  height?: Maybe<Scalars['Float']>;
  /** Length */
  length?: Maybe<Scalars['Float']>;
  /** Mass unit */
  massUnit?: Maybe<MassUnit>;
  /** Weight */
  weight?: Maybe<Scalars['Float']>;
  /** Width */
  width?: Maybe<Scalars['Float']>;
};
// ...
export type SocialMetadata = {
  __typename?: 'SocialMetadata';
  /** Default share message to use when sharing this product on this social network */
  message?: Maybe<Scalars['String']>;
  /** Which social network is this metadata for */
  service?: Maybe<SocialNetwork>;
};
// ...
export type Tag = Deletable &
  Node & {
    __typename?: 'Tag';
    /** The tag ID */
    _id: Scalars['ID'];
    /** The date and time at which this tag was created */
    createdAt: Scalars['DateTime'];
    /** A string of the title to be displayed on a Tag Listing Page */
    displayTitle?: Maybe<Scalars['String']>;
    /** A list of the IDs of top products in this tag */
    featuredProductIds?: Maybe<Array<Maybe<Scalars['ID']>>>;
    /** A string containing the hero image url for a Tag Listing Page */
    heroMediaUrl?: Maybe<Scalars['String']>;
    /**
     * If `true`, this object should be considered deleted. Soft deleted objects are not
     * returned in query results unless you explicitly ask for them.
     */
    isDeleted: Scalars['Boolean'];
    /** If `true`, this tag should be shown at the top level of the tag hierarchy */
    isTopLevel: Scalars['Boolean'];
    /** If `true`, this tag's Tag Listing Page should be visible to the public */
    isVisible: Scalars['Boolean'];
    /** Arbitrary additional metadata about this tag */
    metafields?: Maybe<Array<Maybe<Metafield>>>;
    /** The display name for the tag. This is unique within a given shop. */
    name: Scalars['String'];
    /** The tag's position relative to other tags at the same level of the tag hierarchy */
    position?: Maybe<Scalars['Int']>;
    /** The shop to which this tag belongs */
    shop: Shop;
    /** A unique URL-safe string representing this tag for links */
    slug?: Maybe<Scalars['String']>;
    /** A list of the IDs of tags that have this tag as their parent in the tag hierarchy, in the user-defined order */
    subTagIds: Array<Maybe<Scalars['ID']>>;
    /** A paged list of tags that have this tag as their parent in the tag hierarchy. Currently only three levels are supported. */
    subTags?: Maybe<TagConnection>;
    /** The date and time at which this tag was last updated */
    updatedAt: Scalars['DateTime'];
  };